+++ app/app.go
@@ -8,6 +8,8 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/archway-network/archway/x/genmsg"
+
 	wasmdKeeper "github.com/CosmWasm/wasmd/x/wasm/keeper"
 	wasmdTypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	cosmwasm "github.com/CosmWasm/wasmvm"
@@ -211,6 +213,7 @@ var (
 		wasm.AppModuleBasic{},
 		tracking.AppModuleBasic{},
 		rewards.AppModuleBasic{},
+		genmsg.AppModule{},
 	)
 
 	// module account permissions
@@ -611,6 +614,7 @@ func NewArchwayApp(
 		ibcfee.NewAppModule(app.IBCFeeKeeper),
 		tracking.NewAppModule(app.appCodec, app.TrackingKeeper),
 		rewards.NewAppModule(app.appCodec, app.RewardsKeeper),
+		genmsg.NewAppModule(app.MsgServiceRouter()),
 		crisis.NewAppModule(&app.CrisisKeeper, skipGenesisInvariants), // always be last to make sure that it checks for all invariants and not only part of them
 	)
 
@@ -631,6 +635,7 @@ func NewArchwayApp(
 		govtypes.ModuleName,
 		crisistypes.ModuleName, // doesn't have BeginBlocker, so order is not important
 		genutiltypes.ModuleName,
+		genmsg.ModuleName,
 		authz.ModuleName,
 		feegrant.ModuleName,
 		paramstypes.ModuleName,
@@ -662,6 +667,7 @@ func NewArchwayApp(
 		slashingtypes.ModuleName,
 		minttypes.ModuleName,
 		genutiltypes.ModuleName,
+		genmsg.ModuleName,
 		evidencetypes.ModuleName,
 		paramstypes.ModuleName,
 		upgradetypes.ModuleName,
@@ -707,6 +713,7 @@ func NewArchwayApp(
 		wasm.ModuleName,
 		// wasm gas tracking
 		trackingTypes.ModuleName,
+		genmsg.ModuleName,
 		// invariants checks are always the last to run
 		crisistypes.ModuleName,
 	)